requirements.txt
# Framework
flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Google Cloud
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.16.0

# LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-google-vertexai>=1.0.0

# PDF Generation
weasyprint>=60.0
jinja2>=3.1.0

# HTTP Requests
requests>=2.31.0

# Utilities
pydantic>=2.0.0

